requirements.txt
anthropic>=0.40.0
pyyaml>=6.0
jinja2>=3.1
requests>=2.31
